feat: add tests for splitNumber
