+++ packages/utils/splitNumber.ts
